+++ balancers/balancers_test.go
@@ -15,8 +15,8 @@ import (
 func TestPreferLocalDC(t *testing.T) {
 	conns := []conn.Conn{
 		&mock.Conn{AddrField: "1", LocationField: "1"},
-		&mock.Conn{AddrField: "2", State: state.Online, LocationField: "2"},
-		&mock.Conn{AddrField: "3", State: state.Online, LocationField: "2"},
+		&mock.Conn{AddrField: "2", StateField: state.Online, LocationField: "2"},
+		&mock.Conn{AddrField: "3", StateField: state.Online, LocationField: "2"},
 	}
 	rr := PreferNearestDC(RandomChoice())
 	require.False(t, rr.AllowFallback)
@@ -26,8 +26,8 @@ func TestPreferLocalDC(t *testing.T) {
 func TestPreferLocalDCWithFallBack(t *testing.T) {
 	conns := []conn.Conn{
 		&mock.Conn{AddrField: "1", LocationField: "1"},
-		&mock.Conn{AddrField: "2", State: state.Online, LocationField: "2"},
-		&mock.Conn{AddrField: "3", State: state.Online, LocationField: "2"},
+		&mock.Conn{AddrField: "2", StateField: state.Online, LocationField: "2"},
+		&mock.Conn{AddrField: "3", StateField: state.Online, LocationField: "2"},
 	}
 	rr := PreferNearestDCWithFallBack(RandomChoice())
 	require.True(t, rr.AllowFallback)
@@ -36,9 +36,9 @@ func TestPreferLocalDCWithFallBack(t *testing.T) {
 
 func TestPreferLocations(t *testing.T) {
 	conns := []conn.Conn{
-		&mock.Conn{AddrField: "1", LocationField: "zero", State: state.Online},
-		&mock.Conn{AddrField: "2", State: state.Online, LocationField: "one"},
-		&mock.Conn{AddrField: "3", State: state.Online, LocationField: "two"},
+		&mock.Conn{AddrField: "1", LocationField: "zero", StateField: state.Online},
+		&mock.Conn{AddrField: "2", StateField: state.Online, LocationField: "one"},
+		&mock.Conn{AddrField: "3", StateField: state.Online, LocationField: "two"},
 	}
 
 	rr := PreferLocations(RandomChoice(), "zero", "two")
@@ -48,9 +48,9 @@ func TestPreferLocations(t *testing.T) {
 
 func TestPreferLocationsWithFallback(t *testing.T) {
 	conns := []conn.Conn{
-		&mock.Conn{AddrField: "1", LocationField: "zero", State: state.Online},
-		&mock.Conn{AddrField: "2", State: state.Online, LocationField: "one"},
-		&mock.Conn{AddrField: "3", State: state.Online, LocationField: "two"},
+		&mock.Conn{AddrField: "1", LocationField: "zero", StateField: state.Online},
+		&mock.Conn{AddrField: "2", StateField: state.Online, LocationField: "one"},
+		&mock.Conn{AddrField: "3", StateField: state.Online, LocationField: "two"},
 	}
 
 	rr := PreferLocationsWithFallback(RandomChoice(), "zero", "two")