+++ config/config.go
@@ -23,7 +23,6 @@ type Config struct {
 
 	trace              *trace.Driver
 	dialTimeout        time.Duration
-	connectionTTL      time.Duration
 	balancerConfig     *balancerConfig.Config
 	secure             bool
 	endpoint           string
@@ -74,11 +73,11 @@ func (c *Config) Meta() *meta.Meta {
 	return c.meta
 }
 
-// ConnectionTTL defines interval for parking grpc connections.
-//
-// If ConnectionTTL is zero - connections are not park.
+// Deprecated: connection parking was removed; ConnectionTTL always returns 0.
+// Will be removed after Dec 2026
+// Read about versioning policy: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#deprecated
 func (c *Config) ConnectionTTL() time.Duration {
-	return c.connectionTTL
+	return 0
 }
 
 // Secure is a flag for secure connection
@@ -222,10 +221,11 @@ func WithUserAgent(userAgent string) Option {
 	}
 }
 
+// Deprecated: connection parking was removed; this option is now a no-op.
+// Will be removed after Dec 2026
+// Read about versioning policy: https://github.com/ydb-platform/ydb-go-sdk/blob/master/VERSIONING.md#deprecated
 func WithConnectionTTL(ttl time.Duration) Option {
-	return func(c *Config) {
-		c.connectionTTL = ttl
-	}
+	return func(*Config) {}
 }
 
 func WithCredentials(credentials credentials.Credentials) Option {